chore: refactor plugins into separate files
