+++ packages/vite-plugin-angular/src/lib/angular-vite-plugin.ts
@@ -1,12 +1,10 @@
 import { CompilerHost, NgtscProgram } from '@angular/compiler-cli';
 import { transformAsync } from '@babel/core';
 import angularApplicationPreset from '@angular-devkit/build-angular/src/babel/presets/application';
-import { requiresLinking } from '@angular-devkit/build-angular/src/babel/webpack-loader';
 import * as ts from 'typescript';
 import { ModuleNode, Plugin, PluginContainer, ViteDevServer } from 'vite';
 import { loadEsmModule } from '@angular-devkit/build-angular/src/utils/load-esm';
 import { createJitResourceTransformer } from '@angular-devkit/build-angular/src/builders/browser-esbuild/angular/jit-resource-transformer';
-import { readFileSync } from 'fs';
 import { createCompilerPlugin } from './compiler-plugin';
 import {
   hasStyleUrls,
@@ -15,6 +13,8 @@ import {
   resolveTemplateUrls,
 } from './component-resolvers';
 import { augmentHostWithResources } from './host';
+import { jitPlugin } from './angular-jit-plugin';
+import { buildOptimizerPlugin } from './angular-build-optimizer-plugin';
 
 export interface PluginOptions {
   tsconfig?: string;
@@ -98,15 +98,11 @@ export function angular(options?: PluginOptions): Plugin[] {
   let cssPlugin: Plugin | undefined;
   let styleTransform: PluginContainer['transform'] | undefined;
 
-  return [
-    {
+  function angularPlugin(): Plugin {
+    return {
       name: '@analogjs/vite-plugin-angular',
       async config(config, { command }) {
         watchMode = command === 'serve';
-
-        if (jit) {
-          config.esbuild = false;
-        }
 
         compilerCli = await loadEsmModule<
           typeof import('@angular/compiler-cli')
@@ -114,7 +110,18 @@ export function angular(options?: PluginOptions): Plugin[] {
 
         return {
           optimizeDeps: {
-            include: ['rxjs/operators', 'rxjs'],
+            include: [
+              '@angular/animations',
+              '@angular/common',
+              '@angular/common/http',
+              '@angular/core',
+              '@angular/platform-browser',
+              '@angular/platform-browser/animations',
+              '@angular/platform-browser-dynamic',
+              'rxjs/operators',
+              'rxjs',
+            ],
+            exclude: ['@angular/platform-server'],
             esbuildOptions: {
               plugins: [
                 createCompilerPlugin({
@@ -122,7 +129,7 @@ export function angular(options?: PluginOptions): Plugin[] {
                   sourcemap: !isProd,
                   advancedOptimizations: isProd,
                   jit,
-                }) as any,
+                }),
               ],
               define: {
                 ngJitMode: 'false',
@@ -321,139 +328,18 @@ export function angular(options?: PluginOptions): Plugin[] {
 
         return undefined;
       },
-    },
-    {
-      name: '@analogjs/vite-plugin-angular-jit',
-      resolveId(id: string) {
-        if (id.startsWith('virtual:angular')) {
-          return `\0${id}`;
-        }
+    };
+  }
 
-        return;
-      },
-      async load(id: string) {
-        if (id.includes('virtual:angular:jit:template:file;')) {
-          const contents = readFileSync(id.split('file;')[1], 'utf-8');
-
-          return `export default \`${contents}\`;`;
-        } else if (id.includes('virtual:angular:jit:style:inline;')) {
-          const styleId = id.split('style:inline;')[1];
-
-          const decodedStyles = Buffer.from(
-            decodeURIComponent(styleId),
-            'base64'
-          ).toString();
-
-          let styles: string | undefined = '';
-
-          try {
-            const compiled = await styleTransform!(
-              decodedStyles,
-              `${styleId}.${pluginOptions.inlineStylesExtension}?direct`
-            );
-            styles = compiled?.code;
-          } catch (e) {
-            console.error(`${e}`);
-          }
-
-          return `export default \`${styles}\``;
-        }
-
-        return;
-      },
-    },
-    {
-      name: '@analogjs/vite-plugin-angular-optimizer',
-      apply: 'build',
-      config() {
-        return {
-          esbuild: {
-            legalComments: 'none',
-            keepNames: false,
-            define: isProd
-              ? {
-                  ngDevMode: 'false',
-                  ngJitMode: 'false',
-                  ngI18nClosureMode: 'false',
-                }
-              : undefined,
-            supported: {
-              // Native async/await is not supported with Zone.js. Disabling support here will cause
-              // esbuild to downlevel async/await to a Zone.js supported form.
-              'async-await': false,
-              // Zone.js also does not support async generators or async iterators. However, esbuild does
-              // not currently support downleveling either of them. Instead babel is used within the JS/TS
-              // loader to perform the downlevel transformation. They are both disabled here to allow
-              // esbuild to handle them in the future if support is ever added.
-              // NOTE: If esbuild adds support in the future, the babel support for these can be disabled.
-              'async-generator': false,
-              'for-await': false,
-            },
-          },
-        };
-      },
-      async transform(code, id) {
-        if (/\.[cm]?js$/.test(id)) {
-          const angularPackage = /[\\/]node_modules[\\/]@angular[\\/]/.test(id);
-
-          const linkerPluginCreator = (
-            await loadEsmModule<
-              typeof import('@angular/compiler-cli/linker/babel')
-            >('@angular/compiler-cli/linker/babel')
-          ).createEs2015LinkerPlugin;
-
-          const forceAsyncTransformation =
-            !/[\\/][_f]?esm2015[\\/]/.test(id) &&
-            /for\s+await\s*\(|async\s+function\s*\*/.test(code);
-          const shouldLink = await requiresLinking(id, code);
-          const useInputSourcemap = (!isProd ? undefined : false) as undefined;
-
-          if (!forceAsyncTransformation && !isProd && !shouldLink) {
-            return {
-              code: isProd
-                ? code.replace(/^\/\/# sourceMappingURL=[^\r\n]*/gm, '')
-                : code,
-            };
-          }
-
-          const result = await transformAsync(code, {
-            filename: id,
-            inputSourceMap: useInputSourcemap,
-            sourceMaps: !isProd ? 'inline' : false,
-            compact: false,
-            configFile: false,
-            babelrc: false,
-            browserslistConfigFile: false,
-            plugins: [],
-            presets: [
-              [
-                angularApplicationPreset,
-                {
-                  angularLinker: {
-                    shouldLink,
-                    jitMode: false,
-                    linkerPluginCreator,
-                  },
-                  forceAsyncTransformation,
-                  optimize: isProd && {
-                    looseEnums: angularPackage,
-                    pureTopLevel: angularPackage,
-                  },
-                },
-              ],
-            ],
-          });
-
-          return {
-            code: result?.code || '',
-            map: result?.map as any,
-          };
-        }
-
-        return;
-      },
-    },
-  ];
+  return [
+    angularPlugin(),
+    (jit &&
+      jitPlugin({
+        styleTransform: styleTransform!,
+        inlineStylesExtension: pluginOptions.inlineStylesExtension,
+      })) as Plugin,
+    buildOptimizerPlugin({ isProd }),
+  ].filter(Boolean) as Plugin[];
 
   function setupCompilation() {
     const { options: tsCompilerOptions, rootNames: rn } =
